third_party/WebKit/LayoutTests/accessibility/input-type-range-aria-value-expected.txt

This tests that ARIA attributes can be used to override the value of an input type=range even if it doesn't have an ARIA role of slider, because it has an *implicit* ARIA role of slider already. http://www.w3.org/TR/html-aapi/#el-78

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axRange1.minValue is 1
PASS axRange1.maxValue is 10
PASS axRange1.intValue is 5
PASS axRange2.minValue is 10
PASS axRange2.maxValue is 100
PASS axRange2.intValue is 50
PASS axRange3.valueDescription is 'AXValueDescription: 5 of 10'
PASS successfullyParsed is true

TEST COMPLETE

  
